02_Applications/02_aurex-launchpad/backend/requirements.txt
# ================================================================================
# AUREX LAUNCHPAD™ BACKEND REQUIREMENTS
# Production-ready dependencies for ESG assessment platform
# Created: August 4, 2025
# ================================================================================

# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography>=41.0.0
PyJWT==2.8.0

# HTTP client
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment management
python-dotenv==1.0.0

# Encryption & Security
sqlalchemy-utils==0.41.1

# Date/Time handling
python-dateutil==2.8.2

# Logging & Monitoring
structlog==23.2.0
rich==13.7.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality (development)
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Performance monitoring
psutil==5.9.6
prometheus-client==0.19.0

# ESG-specific libraries
pandas==2.1.3
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.0

# File handling for reports
openpyxl==3.1.2
reportlab==4.0.7
jinja2==3.1.2

# Production optimizations
orjson==3.9.10
uvloop==0.19.0
httptools==0.6.1

# Additional utilities
aiofiles==23.2.1
healthcheck==1.3.3
